+++ src/lib.rs
@@ -45,7 +45,6 @@ At last, the `id` attribute is stored as a separate value and not as part of the
 
 #![forbid(unsafe_code)]
 #![warn(missing_docs)]
-#![recursion_limit="128"] // For error_chain.
 
 #![cfg_attr(feature="clippy", feature(plugin))]
 #![cfg_attr(feature="clippy", plugin(clippy))]
@@ -55,7 +54,7 @@ At last, the `id` attribute is stored as a separate value and not as part of the
 #![cfg_attr(feature="clippy", allow(new_without_default_derive))]
 
 #[macro_use] extern crate log;
-#[macro_use] extern crate error_chain;
+#[macro_use] extern crate failure;
 extern crate svgparser;
 extern crate simplecss;
 extern crate float_cmp;
@@ -90,23 +89,15 @@ mod writer;
 
 pub use attribute::*;
 pub use dom::*;
-pub use error::{
-    Error,
-    ErrorKind,
-};
+pub use error::Error;
 pub use name::*;
 pub use traits::*;
 pub use types::*;
 pub use writer::*;
-
-pub use parser::{
-    ParseOptions,
-};
+pub use parser::ParseOptions;
 
 pub use svgparser::{
     AttributeId,
-    ChainedErrorExt,
     ElementId,
-    ErrorPos,
     ValueId,
 };