+++ src/types/path/parser.rs
@@ -4,9 +4,12 @@
 
 use std::str::FromStr;
 
-use svgparser;
 use svgparser::{
-    Error as ParseError,
+    self,
+    StreamError,
+};
+
+use svgparser::xmlparser::{
     StrSpan,
     FromSpan,
 };
@@ -24,10 +27,8 @@ use super::{
 impl_from_str!(Path);
 
 impl ParseFromSpan for Path {
-    type Err = ParseError;
-
     // TODO: can't fail
-    fn from_span(span: StrSpan) -> Result<Path, ParseError> {
+    fn from_span(span: StrSpan) -> Result<Path, Self::Err> {
         use svgparser::path::Token;
 
         let tokens = svgparser::path::Tokenizer::from_span(span);
@@ -38,59 +39,56 @@ impl ParseFromSpan for Path {
                 Token::MoveTo { abs, x, y } => {
                     Segment {
                         absolute: abs,
-                        data: SegmentData::MoveTo { x: x, y: y },
+                        data: SegmentData::MoveTo { x, y },
                     }
                 }
                 Token::LineTo { abs, x, y } => {
                     Segment {
                         absolute: abs,
-                        data: SegmentData::LineTo { x: x, y: y },
+                        data: SegmentData::LineTo { x, y },
                     }
                 }
                 Token::HorizontalLineTo { abs, x } => {
                     Segment {
                         absolute: abs,
-                        data: SegmentData::HorizontalLineTo { x: x },
+                        data: SegmentData::HorizontalLineTo { x },
                     }
                 }
                 Token::VerticalLineTo { abs, y } => {
                     Segment {
                         absolute: abs,
-                        data: SegmentData::VerticalLineTo { y: y },
+                        data: SegmentData::VerticalLineTo { y },
                     }
                 }
                 Token::CurveTo { abs, x1, y1, x2, y2, x, y } => {
                     Segment {
                         absolute: abs,
-                        data: SegmentData::CurveTo { x1: x1, y1: y1, x2: x2, y2: y2, x: x, y: y },
+                        data: SegmentData::CurveTo { x1, y1, x2, y2, x, y },
                     }
                 }
                 Token::SmoothCurveTo { abs, x2, y2, x, y } => {
                     Segment {
                         absolute: abs,
-                        data: SegmentData::SmoothCurveTo { x2: x2, y2: y2, x: x, y: y },
+                        data: SegmentData::SmoothCurveTo { x2, y2, x, y },
                     }
                 }
                 Token::Quadratic { abs, x1, y1, x, y } => {
                     Segment {
                         absolute: abs,
-                        data: SegmentData::Quadratic { x1: x1, y1: y1, x: x, y: y },
+                        data: SegmentData::Quadratic { x1, y1, x, y },
                     }
                 }
                 Token::SmoothQuadratic { abs, x, y } => {
                     Segment {
                         absolute: abs,
-                        data: SegmentData::SmoothQuadratic { x: x, y: y },
+                        data: SegmentData::SmoothQuadratic { x, y },
                     }
                 }
                 Token::EllipticalArc { abs, rx, ry, x_axis_rotation, large_arc, sweep, x, y } => {
                     Segment {
                         absolute: abs,
                         data: SegmentData::EllipticalArc {
-                            rx: rx, ry: ry,
-                            x_axis_rotation: x_axis_rotation,
-                            large_arc: large_arc, sweep: sweep,
-                            x: x, y: y
+                            rx, ry, x_axis_rotation, large_arc, sweep, x, y
                         },
                     }
                 }