+++ tests/parser.rs
@@ -9,7 +9,6 @@ extern crate simplecss;
 use svgdom::{
     AttributeId as AId,
     AttributeValue,
-    ChainedErrorExt,
     Color,
     Document,
     ElementId as EId,
@@ -39,26 +38,26 @@ macro_rules! test_resave {
 
 #[test]
 fn parse_empty_1() {
-    assert_eq!(Document::from_str("").err().unwrap().full_chain(),
-        "Error: the document does not have any nodes");
+    assert_eq!(Document::from_str("").err().unwrap().to_string(),
+        "the document does not have any nodes");
 }
 
 #[test]
 fn parse_empty_2() {
-    assert_eq!(Document::from_str("\n \t").err().unwrap().full_chain(),
-        "Error: the document does not have any nodes");
+    assert_eq!(Document::from_str("\n \t").err().unwrap().to_string(),
+        "the document does not have any nodes");
 }
 
 #[test]
 fn parse_empty_3() {
-    assert_eq!(Document::from_str("<rect/>").err().unwrap().full_chain(),
-        "Error: the document does not have an SVG element");
+    assert_eq!(Document::from_str("<rect/>").err().unwrap().to_string(),
+        "the document does not have an SVG element");
 }
 
 #[test]
 fn parse_empty_4() {
-    assert_eq!(Document::from_str("<?xml version='1.0'?>").err().unwrap().full_chain(),
-        "Error: the document does not have an SVG element");
+    assert_eq!(Document::from_str("<?xml version='1.0'?>").err().unwrap().to_string(),
+        "the document does not have an SVG element");
 }
 
 #[test]
@@ -289,8 +288,8 @@ fn parse_css_11() {
     </style>
 </svg>");
 
-    assert_eq!(res.err().unwrap().full_chain(),
-        "Error: Unsupported token at 3:9");
+    assert_eq!(res.err().unwrap().to_string(),
+        "Unsupported token at 3:9");
 }
 
 test_resave!(parse_css_12,
@@ -316,8 +315,8 @@ fn parse_css_13() {
     </style>
 </svg>");
 
-    assert_eq!(res.err().unwrap().full_chain(),
-        "Error: unsupported CSS at 3:9");
+    assert_eq!(res.err().unwrap().to_string(),
+        "unsupported CSS at 3:9");
 }
 
 test_resave!(parse_css_14,
@@ -349,8 +348,8 @@ fn parse_css_15() {
     </style>
 </svg>");
 
-    assert_eq!(res.err().unwrap().full_chain(),
-        "Error: unsupported CSS at 3:10");
+    assert_eq!(res.err().unwrap().to_string(),
+        "unsupported CSS at 3:10");
 }
 
 #[test]
@@ -363,8 +362,8 @@ fn parse_css_16() {
     </style>
 </svg>");
 
-    assert_eq!(res.err().unwrap().full_chain(),
-        "Error: unsupported CSS at 3:10");
+    assert_eq!(res.err().unwrap().to_string(),
+        "unsupported CSS at 3:10");
 }
 
 // empty style
@@ -614,8 +613,8 @@ fn parse_iri_with_fallback_3() {
     <rect fill='url(#rg1) none'/>
 </svg>");
 
-    assert_eq!(doc.err().unwrap().full_chain(),
-               "Error: valid FuncIRI(#rg1) with fallback value is not supported");
+    assert_eq!(doc.err().unwrap().to_string(),
+               "valid FuncIRI(#rg1) with fallback value is not supported");
 }
 
 #[test]
@@ -628,8 +627,8 @@ fn parse_iri_with_fallback_4() {
     <radialGradient id='rg1'/>
 </svg>");
 
-    assert_eq!(doc.err().unwrap().full_chain(),
-               "Error: valid FuncIRI(#rg1) with fallback value is not supported");
+    assert_eq!(doc.err().unwrap().to_string(),
+               "valid FuncIRI(#rg1) with fallback value is not supported");
 }
 
 test_resave!(parse_filter_iri_1,
@@ -696,8 +695,8 @@ fn parse_entity_5() {
     <!ENTITY Viewport1 \"<rect/>\">
 ]>
 <svg fill='&st1;'/>");
-    assert_eq!(doc.err().unwrap().full_chain(),
-               "Error: unsupported ENTITY data at 2:25");
+    assert_eq!(doc.err().unwrap().to_string(),
+               "unsupported ENTITY data at 2:25");
 }
 
 #[test]
@@ -707,8 +706,8 @@ fn parse_entity_6() {
     <!ENTITY Viewport1 \" \t\n<rect/>\">
 ]>
 <svg fill='&st1;'/>");
-    assert_eq!(doc.err().unwrap().full_chain(),
-               "Error: unsupported ENTITY data at 2:25");
+    assert_eq!(doc.err().unwrap().to_string(),
+               "unsupported ENTITY data at 2:25");
 }
 
 test_resave!(skip_unknown_refs_1,